Split online tests for BioSQL
We need internet access for NCBI Entrez to load taxonomy
information automatically.

Based on a pull request from @ctSkennerton which was updated
by @peterjc following changes on the master branch, with some
painful workarounds for global BioSQL configuration variables.
