+++ Tests/test_BioSQL_sqlite3_online.py
@@ -4,22 +4,22 @@
 # as part of this package.
 
 """Run BioSQL tests using SQLite"""
-from Bio import MissingExternalDependencyError
-from BioSQL import BioSeqDatabase
 
-from common_BioSQL import *
+from common_BioSQL import temp_db_filename
+from common_BioSQL_online import *
 
-DBDRIVER = 'mysql.connector'
-DBTYPE = 'mysql'
+# Constants for the database driver
+DBDRIVER = 'sqlite3'
+DBTYPE = 'sqlite'
 
-DBHOST, DBUSER, DBPASSWD, TESTDB = load_biosql_ini(DBTYPE)
+DBHOST = None
+DBUSER = 'root'
+DBPASSWD = None
+TESTDB = temp_db_filename()
 
 # This will abort if driver not installed etc:
-TESTDB = check_config(DBDRIVER, DBTYPE, DBHOST, DBUSER, DBPASSWD, TESTDB)
-
-# Some of the unit tests don't create their own database,
-# so just in case there is no database already:
-create_database()
+check_config(DBDRIVER, DBTYPE, DBHOST, DBUSER, DBPASSWD, TESTDB)
+share_config(DBDRIVER, DBTYPE, DBHOST, DBUSER, DBPASSWD, TESTDB)
 
 if __name__ == "__main__":
     # Run the test cases
